requirements.txt
bs4
python-telegram-bot
